rename overview widgets
