+++ src/widgets/allergies/allergy-card-level-two.test.tsx
@@ -3,7 +3,7 @@ import { performPatientAllergySearch } from "./allergy-intolerance.resource";
 import { render, cleanup, wait } from "@testing-library/react";
 import { act } from "react-dom/test-utils";
 import { BrowserRouter, match } from "react-router-dom";
-import { AllergyCardLevelTwo } from "./allergy-card-level-two.component";
+import { AllergyOverviewLevelTwo } from "./allergy-card-level-two.component";
 import { useCurrentPatient } from "../../../__mocks__/openmrs-esm-api.mock";
 
 const mockPerformPatientAllergySearch = performPatientAllergySearch as jest.Mock;
@@ -292,7 +292,7 @@ describe("AlleryCardLevelTwo />", () => {
     );
     wrapper = render(
       <BrowserRouter>
-        <AllergyCardLevelTwo match={match} />
+        <AllergyOverviewLevelTwo match={match} />
       </BrowserRouter>
     );
 
@@ -308,7 +308,7 @@ describe("AlleryCardLevelTwo />", () => {
     );
     wrapper = render(
       <BrowserRouter>
-        <AllergyCardLevelTwo match={match} />
+        <AllergyOverviewLevelTwo match={match} />
       </BrowserRouter>
     );
 
@@ -325,7 +325,7 @@ describe("AlleryCardLevelTwo />", () => {
     );
     wrapper = render(
       <BrowserRouter>
-        <AllergyCardLevelTwo match={match} />
+        <AllergyOverviewLevelTwo match={match} />
       </BrowserRouter>
     );
