+++ src/widgets/allergies/allergy-overview.css
