+++ src/widgets/conditions/conditions-overview.css
