+++ src/widgets/medications/medications-overview.css
