+++ src/widgets/notes/notes-overview.css
