+++ src/widgets/vitals/vitals-overview.css
